editor functionality and documentation
Added tons of documentation which can also be viewed with the Godot autodocs. Fixed SoundManager not playing sounds lol and Level not playing music or ambience. All kinds of stupid things fixed on the ResourceData resources and Global.gd in the examples. Added some of my own music for examples, too.